Add files via upload
The paper is to abolish ragging activities as ragging is one of the most common problems that have a deep root in the Education Society.
Although there are many laws against ragging in India still this practice prevails in most of the institutions of higher studies.
In this paper, we have proposed an efficient computer vision based framework to detect ragging/violence during the actual time of occurrence of the event. The framework of the system is based on motion analysis.